issue #89: refactored authentication and put login and register on same page (work commit)
