+++ frontend/src/App.tsx
@@ -1,8 +1,7 @@
 import * as React                            from 'react';
 import 'antd/dist/antd.css';
 import {HashRouter as Router, Route, Switch} from 'react-router-dom';
-import Register                              from './components/register/Register';
-import Login                                 from './components/login/Login';
+import Authentication                        from './components/login/Authentication';
 import {configureI18N}                       from './translations/translations';
 import Home                                  from './components/home/Home';
 
@@ -13,8 +12,7 @@ export default class App extends React.Component {
         return (
             <Router>
                 <Switch>
-                    <Route path='/register' component={Register}/>
-                    <Route path='/login' component={Login}/>
+                    <Route path='/authentication' component={Authentication}/>
                     <Route path='/' component={Home}/>
                 </Switch>
             </Router>
