+++ frontend/src/translations/de_DE/translations.json
@@ -19,5 +19,7 @@
     "UnauthorizedOperation": "Sie sind nicht berechtigt, diese Aktion durchzuführen. Wenn Sie diese Aktion trotzdem durchführen möchten, wenden Sie sich bitte an den Administrator, um die Erlaubnis einzuholen.",
     "NotFound": "Die angefragte Ressource ist nicht verfügbar. "
   },
-  "logout": "Ausloggen"
+  "logout": "Ausloggen",
+  "surname": "Nachname",
+  "firstName": "Vorname"
 }
